buffetology/data_fetchers/ft_fetcher.py
import requests
import pandas as pd
from bs4 import BeautifulSoup
from typing import Dict, List
from .base_fetcher import BaseDataFetcher

class FinancialTimesFetcher(BaseDataFetcher):
    def __init__(self, config_path: str = "buffetology/config/config.yaml"):
        super().__init__(config_path)
        self.username = self.config['data_provider']['financial_times']['username']
        self.password = self.config['data_provider']['financial_times']['password']
        self.base_url = "https://markets.ft.com"
        self.session = self._create_session()
    
    def _create_session(self) -> requests.Session:
        session = requests.Session()
        login_url = f"{self.base_url}/login"
        login_data = {
            "username": self.username,
            "password": self.password
        }
        response = session.post(login_url, data=login_data)
        response.raise_for_status()
        return session
    
    def get_financial_statements(self, ticker: str) -> Dict[str, pd.DataFrame]:
        cache_path = self._get_cache_path(ticker, "financial_statements")
        cached_data = self._load_from_cache(cache_path)
        if cached_data is not None:
            return {
                "income": cached_data[cached_data['statement'] == 'income'],
                "balance": cached_data[cached_data['statement'] == 'balance'],
                "cashflow": cached_data[cached_data['statement'] == 'cashflow']
            }
        
        # Fetch financial statements from FT
        url = f"{self.base_url}/data/equities/tearsheet/financials?s={ticker}"
        response = self.session.get(url)
        response.raise_for_status()
        
        # Parse the HTML response
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Extract financial data (this is a simplified example)
        # In a real implementation, you would need to parse the specific HTML structure
        income = self._parse_financial_table(soup, "income")
        balance = self._parse_financial_table(soup, "balance")
        cashflow = self._parse_financial_table(soup, "cashflow")
        
        # Add statement type for caching
        income['statement'] = 'income'
        balance['statement'] = 'balance'
        cashflow['statement'] = 'cashflow'
        
        # Combine for caching
        combined = pd.concat([income, balance, cashflow])
        self._save_to_cache(combined, cache_path)
        
        return {
            "income": income,
            "balance": balance,
            "cashflow": cashflow
        }
    
    def _parse_financial_table(self, soup: BeautifulSoup, statement_type: str) -> pd.DataFrame:
        # This is a placeholder implementation
        # In a real implementation, you would parse the specific HTML structure
        # of the Financial Times financial statements
        return pd.DataFrame()
    
    def get_key_metrics(self, ticker: str) -> pd.DataFrame:
        cache_path = self._get_cache_path(ticker, "key_metrics")
        cached_data = self._load_from_cache(cache_path)
        if cached_data is not None:
            return cached_data
        
        # Fetch key metrics from FT
        url = f"{self.base_url}/data/equities/tearsheet/summary?s={ticker}"
        response = self.session.get(url)
        response.raise_for_status()
        
        # Parse the HTML response
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Extract key metrics (this is a simplified example)
        metrics = self._parse_key_metrics(soup)
        
        self._save_to_cache(metrics, cache_path)
        return metrics
    
    def _parse_key_metrics(self, soup: BeautifulSoup) -> pd.DataFrame:
        # This is a placeholder implementation
        # In a real implementation, you would parse the specific HTML structure
        # of the Financial Times key metrics
        return pd.DataFrame()
    
    def get_stock_price(self, ticker: str, start_date: str, end_date: str) -> pd.DataFrame:
        cache_path = self._get_cache_path(ticker, f"price_{start_date}_{end_date}")
        cached_data = self._load_from_cache(cache_path)
        if cached_data is not None:
            return cached_data
        
        # Fetch historical price data from FT
        url = f"{self.base_url}/data/equities/tearsheet/historical?s={ticker}"
        response = self.session.get(url)
        response.raise_for_status()
        
        # Parse the HTML response
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Extract historical prices (this is a simplified example)
        hist = self._parse_historical_prices(soup)
        
        self._save_to_cache(hist, cache_path)
        return hist
    
    def _parse_historical_prices(self, soup: BeautifulSoup) -> pd.DataFrame:
        # This is a placeholder implementation
        # In a real implementation, you would parse the specific HTML structure
        # of the Financial Times historical prices
        return pd.DataFrame()
    
    def get_sp500_tickers(self, limit: int) -> List[str]:
        cache_path = self._get_cache_path("sp500", "tickers")
        cached_data = self._load_from_cache(cache_path)
        if cached_data is not None:
            return cached_data['ticker'].head(limit).tolist()
        
        # Fetch S&P 500 components from FT
        url = f"{self.base_url}/data/indices/tearsheet/constituents?s=INX:IOM"
        response = self.session.get(url)
        response.raise_for_status()
        
        # Parse the HTML response
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Extract tickers (this is a simplified example)
        tickers = self._parse_sp500_tickers(soup)
        
        # Cache the results
        ticker_df = pd.DataFrame({'ticker': tickers})
        self._save_to_cache(ticker_df, cache_path)
        
        return tickers[:limit]
    
    def _parse_sp500_tickers(self, soup: BeautifulSoup) -> List[str]:
        # This is a placeholder implementation
        # In a real implementation, you would parse the specific HTML structure
        # of the Financial Times S&P 500 constituents
        return [] 